added libraries to include on every header file for documentation. Moved mythrow to seperate file and moved main to cMchecker.cpp. Deleted cMchecker_body.hpp
