+++ include/MultivariateModel.hpp
@@ -1,10 +1,12 @@
 #ifndef MULTIVARIATEMODEL_H
 #define MULTIVARIATEMODEL_H
 
+#include "includeLibs.hpp"
 #include "DataSet.hpp"
 #include "Genealogy.hpp"
 #include "TrialsOutput.hpp"
 #include "globals.hpp"
+#include "myThrow.hpp"
 
 struct MultivariateModel {
 	// model brings together a genealogy, a set of trials (theoretical), and a set of data (measured)
