+++ include/lenFixedCharArray.hpp
@@ -1,6 +1,8 @@
 #ifndef LENFIXEDCHARARRAY_H
 #define LENFIXEDCHARARRAY_H
 
+#include "includeLibs.hpp"
+
 Int lenfixedchararray(char **arr) {
 	Int maxlen = 100; // stop runaways
 	Int len = 0;
